Scatter plot - Perception of corruption vs Ladder score
Ladder Score - This is called the Cantril ladder: it asks respondents to think of a ladder, with the best possible life for them being a 10, and the worst possible life being a 0. They are then asked to rate their own current lives on that 0 to 10 scale
Perception of Corruption - The Corruption Perceptions Index is an index published annually by Berlin-based Transparency International since 1995 which ranks countries "by their perceived levels of public sector corruption, as determined by expert assessments and opinion surveys.